Fix NullPointerException in deleteTabStateFiles
In a small number of cases listFiles() is returning null.
Add in some safeguards to avoid NullPointerException.

Bug: b:342257010
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5564755
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1307544}

+++ chrome/browser/tabpersistence/android/java/src/org/chromium/chrome/browser/tabpersistence/TabStateFileManager.java
@@ -648,7 +648,11 @@ public class TabStateFileManager {
                 });
     }
 
-    private static void deleteFlatBufferFiles(File stateDirectory) {
+    @VisibleForTesting
+    protected static void deleteFlatBufferFiles(File stateDirectory) {
+        if (stateDirectory == null || stateDirectory.listFiles() == null) {
+            return;
+        }
         for (File file : stateDirectory.listFiles()) {
             if (file.getName().startsWith(FLATBUFFER_PREFIX) && !file.delete()) {
                 Log.e(TAG, "Failed to delete FlatBuffer TabState: " + file);
